app/requirements/base.txt
coreapi==2.3.0
Django>=2.1.3, <2.2.0
django-cors-headers>=2.5.2, <2.6.0
django-debug-toolbar==1.11
djangorestframework>=3.9.0, <3.10.0
Markdown==3.1
Pillow>=5.4.1, <5.5.0
flake8>=3.6.0, <3.7.0
pycodestyle>=2.4.0, <2.5.0
Pygments>=2.3.0, <2.4.0
# uWSGI==2.0.18
django-taggit==1.1.0
django-mysql==3.2.0
django-extra-fields==1.2.3